sysd: Split utils module in multiple files
